Set default local port to 3000
Previously, `heroku local web` had to be used for local testing to
ensure `process.env.PORT` was set correctly. Now, the port defaults to
`3000` if `process.env.PORT` is not set. This means that local testing
once again works with `node server.js` as well as `heroku local web`.

Local testing can now be run with:
* `npm start`/`npm run start`: local node server
* `npm run heroku`: local Heroku server (runs node too of course)